requirements.txt
Django==5.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
Pillow==10.4.0
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
django-extensions==3.2.3
python-decouple==3.8
django-filter==24.3
celery==5.3.4
redis==5.0.8
gunicorn==23.0.0
whitenoise==6.9.0